Get the dashboard to render when logged out
